Refactoring BridgeConstants Classes from RSKj
- Changing BridgeConstants classes (regtest, testnet, mainnet, devnet) imports from co.rsk.config package to co.rsk.peg package.
- Doing the same inside the test section.
