+++ src/test/java/co/rsk/federate/adapter/ThinConverterTest.java
@@ -4,7 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
-import co.rsk.config.BridgeRegTestConstants;
+import co.rsk.peg.constants.BridgeRegTestConstants;
 import java.math.BigInteger;
 import java.util.ArrayList;
 import org.junit.jupiter.api.Test;
